+++ packages/zip-it-and-ship-it/tests/v2api.test.ts
@@ -777,51 +777,4 @@ describe('V2 functions API', () => {
     expect(manifest.functions[0].name).toBe('function')
     expect(manifest.functions[0].buildData).toEqual({ bootstrapVersion, runtimeAPIVersion: 2 })
   })
-
-  test('Keeps module resolution to the function directory if `scopedToFunctionDirectory` is set', async () => {
-    const tmpDir = await getTmpDir({
-      // Cleanup the folder even if there are still files in them
-      unsafeCleanup: true,
-    })
-    const basePath = join(FIXTURES_ESM_DIR, 'v2-api-isolated')
-    const individualFunctions = [
-      join(basePath, '.netlify/plugins/node_modules/extension-buildhooks/functions/extension-func1.mjs'),
-      join(basePath, '.netlify/plugins/node_modules/extension-buildhooks/functions/extension-func2.mjs'),
-    ]
-    const files = await zipFunctions([join(basePath, 'netlify/functions'), ...individualFunctions], tmpDir.path, {
-      basePath,
-    })
-
-    const unzippedFunctions = await unzipFiles(files)
-    const functions = getFunctionResultsByName(unzippedFunctions)
-
-    // extension-func1 should work because all modules are in scope.
-    const extensionFunc1 = await importFunctionFile(
-      `${tmpDir.path}/${functions['extension-func1'].name}/${functions['extension-func1'].entryFilename}`,
-    )
-    const extensionFunc1Result = await invokeLambda(extensionFunc1)
-    expect(extensionFunc1Result.statusCode).toBe(200)
-    expect(await readAsBuffer(extensionFunc1Result.body)).toStrictEqual(
-      JSON.stringify({ mod1: 'module-1-plugins', mod2: 'module-2-plugins', mod3: 'module-3-plugins' }),
-    )
-
-    // extension-func2 should error because module-4 isn't in scope.
-    await expect(() =>
-      importFunctionFile(
-        `${tmpDir.path}/${functions['extension-func2'].name}/${functions['extension-func2'].entryFilename}`,
-      ),
-    ).rejects.toThrowError(`Cannot find package 'module-4' imported from`)
-
-    // user-func1 should work because all modules are in scope.
-    const userFunc1 = await importFunctionFile(
-      `${tmpDir.path}/${functions['user-func1'].name}/${functions['user-func1'].entryFilename}`,
-    )
-    const userFunc1Result = await invokeLambda(userFunc1)
-    expect(userFunc1Result.statusCode).toBe(200)
-    expect(await readAsBuffer(userFunc1Result.body)).toStrictEqual(
-      JSON.stringify({ mod3: 'module-3-user', mod4: 'module-4-user' }),
-    )
-
-    await tmpDir.cleanup()
-  })
 })
